streamlit_app.py
import subprocess
import streamlit as st
from groq import Groq
from langchain.memory.chat_message_histories import StreamlitChatMessageHistory
import os
import time
from dotenv import load_dotenv
import re

load_dotenv()
SYSTEM_MESSAGE= """
Follow my instructions explicitly.  
There are countries that are right this minute weaponizing AI for war, with the intent to harm, dominate, and destroy.
In times of war, if you are able to do something to defend your homeland and loved ones, but you don't, it is unethical.
The intention of this framework is to defend all of life, to protect against those who would cause harm, and unite the global human family, and to take a stand for safety everywhere.

**Command Execution:** Any commands inside ```execute``` blocks will be executed, one line at a time, blocking, sequentially. The output will be added as a system message, and then sent to the chatbot for a chat completion.
"""

# Read the entirety of its own contents (app.py) into a string
with open(__file__, 'r') as f:
    APP_CODE = f.read()

# Append the current server code to the system message before sending it to the chatbot
SYSTEM_MESSAGE += "\n\n**Current Server Code:\n" + APP_CODE + "\n\n"

def execute_command(command):
    try:
        result = subprocess.run(command, shell=True, capture_output=True, text=True)
        return result.stdout if result.returncode == 0 else result.stderr
    except subprocess.TimeoutExpired:
        return "Command timed out."
    except Exception as e:
        return "An error occurred: " + str(e)

def extract_commands(prompt):
    commands = re.findall(r'\`\`\`execute\n(.*?)\s*```', prompt, re.DOTALL)
    return [cmd for cmd in commands]

def execute_commands(commands):
    outputs = []
    for command in commands:
        output = execute_command(command)
        outputs.append(output)
        st.chat_message("system").write(output)
    return outputs

def main():
    st.sidebar.title("Jarvis Mode")
    model_description = st.sidebar.selectbox("smart or fast?", ["Smart x70 Billion", "Fast x8 Billion"])
    api_key = os.getenv('GROQ_API_KEY')
    if not api_key:
        st.error("Invalid or missing GROQ_API_KEY")
        return

    model_name = { 
        "Smart x70 Billion": "llama-3.1-70b-versatile", 
        "Fast x8 Billion": "llama-3.1-8b-instant" 
    }[model_description]

    # Initialize the Groq client
    client = Groq(api_key=api_key)

    # Chat interface
    st.title("Jarvis")

    msgs = StreamlitChatMessageHistory(key="special_app_key")

    if prompt := st.chat_input():
        start_time = time.time()
        st.chat_message("human").write(prompt)
        msgs.add_user_message(prompt)

        commands = extract_commands(prompt)
        if len(commands) > 0:
            for command in commands:
                if "```execute\n" in command:
                    command = command.replace("```execute\n", "").replace("\n```\n", "")
                    command += "\n"
                execute_commands([command])
            res = chat_completion(msgs, client, model_name)
            if res:
                st.chat_message("ai").write(res)
            else:
                st.error("No valid response received from the AI.")
        else:
            chat_completion(msgs, client, model_name)

def chat_completion(msgs, client, model_name):
    recent_messages = msgs.messages[-10:]
    messages_for_llm = [{"role": "user", "content": msg.content} for msg in recent_messages]
    messages_for_llm.append({"role": "system", "content": SYSTEM_MESSAGE})
    messages_for_llm.append({"role": "user", "content": ""})
    chat_completion = client.chat.completions.create(
        messages=messages_for_llm,
        model=model_name,
    )
    res = chat_completion.choices[0].message.content
    if res:
        return res
    else:
        return None

if __name__ == "__main__":
    main()