requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.0.4
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
groq==0.10.0
h11==0.14.0
hf_transfer==0.1.8
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
installer==0.7.0
ipython==8.27.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-text-splitters==0.2.2
langsmith==0.1.108
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.4.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.6.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-dotenv==0.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.20.0
ruff==0.6.3
SecretStorage==3.3.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spaces==0.29.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
streamlit==1.33.0
tenacity==8.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trove-classifiers==2024.7.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
watchdog==4.0.2
wcwidth==0.2.13
websockets==12.0
xxhash==3.5.0
yarl==1.9.6
zipp==3.20.1
